create tags.factories_test
