+++ backend-project/small_eod/tags/tests.py
@@ -1,18 +1,6 @@
 from django.test import TestCase
 
 from .factories import TagFactory, TagNamespaceFactory
-from .models import Tag, TagNamespace
-from ..generic.tests import FactoryCreateObjectsMixin
-
-
-class TagFactoryTestCase(FactoryCreateObjectsMixin, TestCase):
-    MODEL = Tag
-    FACTORY = TagFactory
-
-
-class TagNamespaceFactoryTestCase(FactoryCreateObjectsMixin, TestCase):
-    MODEL = TagNamespace
-    FACTORY = TagNamespaceFactory
 
 
 class TagsTestCase(TestCase):
